shares: Fix local file paths for shares
When a share is created and accessed via Public User the scope is set to
"/does/not/exist" which leads to errors as scope is used in recreating
the file path for the share. This commit sets it to default so that the
file lookup points to the right location in stoarge. Some additional
cleanup on variable redeclaration. Tested via creating, accessing and
downloading a share.