Update fr.json
Improve the French translation by fixing three translation errors.

"Upload only" is now translated by "Importation uniquement" instead of "Téléchargement uniquement". Téléchargement was ambigous as it can mean both download and upload and is more often used in the sense of download.
I would have personnaly translated upload by "téléversement" instead of importation as it is even clearer but since importer/importation is already used elsewhere it would have been bad for consistency.

"Save" is now translated by "Sauvegarder" instead of "Economiser". Economiser mean to save money, to economize so it is wrong in this context.

"Clear completed" is now translated by "Effacer les importations achevées" instead of "Effacement terminé". "Effacement terminé" means it already has been cleared which is wrong since it's a button to actually perform the action of cleaning.
The new translation is little long but it's really not possible to make it shorter in French.